+++ CHANGELOG.md
@@ -630,3 +630,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Default prefix from MaskedInput component.
+
+## [1.3.1] - 2020-05-25
+### Added
+- Default margin if exists any SVG in Button component.
+
+### Change
+- Font-family from Button component.
+- Font-weight from Filter component.
+
+## [1.3.2] - 2020-05-25
+### Change
+- Move default margin of SVG to props in Button component.
+
+## [1.4.0] - 2020-05-26
+### Add
+- Props **async** in Select to make the component async.
+- CSS improvements on Editor component.
+
+## [1.4.1] - 2020-05-26
+### Changed
+- Default props **sync** from Select should be false.